+++ Demo/Sources/Demo/DemoViews/ComponentDemoViews.swift
@@ -50,7 +50,7 @@ public enum ComponentDemoViews: String, DemoViews {
     case priming
     case footerButtonView
     case checkmarkTitleView
-    case viewingsView
+    case viewingsListView
     case selfDeclarationView
     case collapsibleContentView
     case columnListsView
@@ -198,8 +198,8 @@ public enum ComponentDemoViews: String, DemoViews {
             return DemoViewController<FooterButtonDemoView>(constrainToBottomSafeArea: false)
         case .checkmarkTitleView:
             return DemoViewController<CheckmarkTitleViewDemoView>()
-        case .viewingsView:
-            return DemoViewController<ViewingsDemoView>()
+        case .viewingsListView:
+            return DemoViewController<ViewingsListDemoView>()
         case .selfDeclarationView:
             return DemoViewController<SelfDeclarationDemoView>()
         case .collapsibleContentView:
